refactor chat interface
